requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# HTML Parsing
beautifulsoup4==4.12.3
lxml==5.1.0

# Logging
structlog==24.1.0

# Validation
pydantic>=2.10.0

# OAuth
authlib==1.3.0
